refactor minor change
